fade in/out, object activate/deactivate command updated
fade in/out, object activate/deactivate command updated
fade in/out, object  활성/비활성화 yarn command 업데이트
