+++ LICENSE.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8b7c954cb45f378418aa8c12180dcb85
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
