+++ README.md.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4e003d495e50d0644936b38d9ce4e5ec
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
