--- custum_yarn_command/custum_yarn_command.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 990be7f7d4c674946886866c0c517f17
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
